add add fee plugin
